include/utime: un-inline iostream code to reduce header dependencies
